plots

+++ src/derpinns/solution.py
@@ -13,7 +13,7 @@ def compare_with_mc(
     params,
     n_prices: int,
     n_simulations: int,
-    dtype=torch.float32,
+    dtype=torch.float64,
     device=torch.device("cpu"),
     seed=None,
     split_output=False,
@@ -32,8 +32,7 @@ def compare_with_mc(
     sigma = torch.as_tensor(params.sigma, dtype=dtype, device=device)
 
     x0 = torch.empty((n_prices, params.n_assets),
-                     dtype=dtype, device=device).uniform_(params.x_min,
-                                                          params.x_max)
+                     dtype=dtype, device=device).uniform_(params.x_min, params.x_max)
     s0 = torch.exp(x0) * params.strike
 
     # not implmeneted in mps, use numpy cholesky instead
@@ -48,21 +47,16 @@ def compare_with_mc(
     vol_sqrt = sigma * np.sqrt(params.tau)
 
     t0 = time.perf_counter()
-    inp = torch.cat([x0, torch.full((n_prices, 1), params.tau,
-                                    dtype=dtype, device=device)], dim=1)
+    inp = torch.cat([x0, torch.full((n_prices, 1), params.tau, dtype=dtype, device=device)], dim=1)
     if split_output:
         # split the output, the net returns a matrix where the first
         # column is the price and next following columns are the first order derivatives of u wrt inputs
-        nn_prices = torch.split(model(inp), 1, dim=1)[
-            0].squeeze() * params.strike
+        nn_prices = torch.split(model(inp), 1, dim=1)[0].squeeze() * params.strike
     else:
         nn_prices = model(inp).squeeze() * params.strike   # (n_prices,)
     nn_time = time.perf_counter() - t0
-
     t0 = time.perf_counter()
-
-    Z = random_samples(n_prices * n_simulations, params.n_assets,
-                       sampler="Sobol")
+    Z = random_samples(n_prices * n_simulations, params.n_assets, sampler="Sobol")
     Z = torch.as_tensor(norm.ppf(Z), dtype=dtype, device=device)
 
     # apply correlation
